use Memory SSA in Dependence Analysis
Use the Memory SSA representation to check whether the base address of two
memory accesses are defined by a same memory SSA node.

This fixes https://github.com/llvm/llvm-project/issues/41488
"Aliasing checks in DA are incorrect", and fixes the first testcase in
https://github.com/llvm/llvm-project/issues/53942
"[DependenceAnalysis] assumes base pointer invariance".

This patch depends on a series of patches to preserve MemorySSA in the passes
that require array dependence analysis. For instance, loop-fuse fails 2 tests
in test/Transforms/LoopFusion because the transform invalidates the MemorySSA.